refactor: dictknife.cliutils
